// src/generated/model/boutgifterFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface BoutgifterFrontend {
    bekreftelse?: boolean;
    husleie: boolean;
    strom: boolean;
    kommunalAvgift: boolean;
    oppvarming: boolean;
    boliglan: boolean;
    annet: boolean;
    skalViseInfoVedBekreftelse: boolean;
}